Match SpiritDoorLock_update 96.03->100% (drop spurious (s16) extsh on rotZ store #20)

+++ src/main/dll/dll_0167_spiritdoorlock.c
@@ -183,7 +183,7 @@ void SpiritDoorLock_update(int obj)
                 }
                 ((GameObject*)obj)->anim.rootMotionScale *= gSpiritDoorLockScaleDecay;
                 ((GameObject*)obj)->anim.rotZ =
-                    (s16)(s32)((f32)(int)((GameObject*)obj)->anim.rotZ - lbl_803E4450 * timeDelta);
+                    (s32)((f32)(int)((GameObject*)obj)->anim.rotZ - lbl_803E4450 * timeDelta);
             }
             else
             {
